ironing out migration to cs kinks
